alteracao trabalho

+++ questao07.py
@@ -1,7 +1,7 @@
 nome=input("nome do aluno: ")
 diciplina=input("qual a diciplina: ")
-nota1=int(input("nota parcial: "))
-nota2=int(input("nota bimestral: "))
+nota1=float(input("nota parcial: "))
+nota2=float(input("nota bimestral: "))
 media=(nota1+nota2)/2
 if media>5:
     s=("aprovado")
